+++ src/flat_vector.cpp
@@ -1,6 +1,6 @@
 #include "../include/flat_vector.h"
 
-//FlatVector::FlatVector(float _x, float _y) {
+//FlatVector::FlatVector(double _x, double _y) {
 	//x = _x;
 	//y = _y;
 //};
@@ -13,20 +13,20 @@ FlatVector FlatVector::operator-(FlatVector const& obj) {
 	return FlatVector(x - obj.x, y - obj.y);
 }
 
-float FlatVector::operator*(FlatVector const& obj) {
+double FlatVector::operator*(FlatVector const& obj) {
 	return x * obj.x + y * obj.y;
 }
 
-FlatVector FlatVector::operator*(float s) {
+FlatVector FlatVector::operator*(double s) {
 	return FlatVector(x * s, y * s);
 }
 
-float FlatVector::length() {
+double FlatVector::length() {
 	return sqrt(x * x + y * y);
 }
 
 FlatVector FlatVector::norm() {
-	float length = this->length();
+	double length = this->length();
 	if (length == 0) {
 		return FlatVector(0,0);
 	}
